pixel-forge/backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# AI Services
google-generativeai==0.3.0
fal-client==0.7.0
elevenlabs==0.2.0

# Image processing
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.0

# Database & Cache
redis==5.0.1
sqlalchemy==2.0.0
asyncpg==0.29.0
alembic==1.12.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.0
aiofiles==23.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development & Testing
pytest==7.4.0
pytest-asyncio==0.21.0
black==23.11.0
ruff==0.1.0